Made TLS support an optional feature, only enabled when configured with --enable-tls
