+++ t/binary.t
@@ -443,7 +443,7 @@ $mc->silent_mutation(::CMD_ADDQ, 'silentadd', 'silentaddval');
     is(1024, $stats{'maxconns'});
     # we run SSL tests over TCP; hence the domain_socket
     # is expected to be NULL.
-    if ($ENV{SSL_TEST}) {
+    if (enabled_tls_testing()) {
         is('NULL', $stats{'domain_socket'});
     } else {
         isnt('NULL', $stats{'domain_socket'});
