+++ t/idle-timeout.t
@@ -32,7 +32,7 @@ mem_stats($sock);   # Network activity, so socket code will see dead socket
 sleep(1);
 # we run SSL tests over TCP; hence IO::Socket::SSL returns
 # '' upon disconnecting with the server.
-if ($ENV{SSL_TEST}) {
+if (enabled_tls_testing()) {
     is($sock->connected(),'', "check disconnected");
 } else {
     is($sock->connected(),undef, "check disconnected");
